updated lineitem schema

+++ micro-benchmarks/aggregation/local/config.h
@@ -82,8 +82,8 @@ using Buffer = EvictionBuffer<PageHashtable, BlockAlloc>;
 using StorageGlobal = PartitionBuffer<PageHashtable, true>;
 
 // #define SCHEMA GRP_KEYS, u32, u32, std::array<char, 4>
-//  TPCH lineitem (not exact)
-#define SCHEMA GRP_KEYS, u64, u64, u64, double, double, double, double, char, char, u64, u64, u64, std::array<char, 25>, std::array<char, 10>, std::array<char, 44>
+//  TPCH lineitem
+#define SCHEMA GRP_KEYS, u64, u64, u64, double, double, double, double, char, char, s32, s32, s32, std::array<char, 25>, std::array<char, 10>, std::array<char, 44>
 // #define SCHEMA GRP_KEYS
 
 using PageTable = PageLocal<SCHEMA>;
